Clean Architecture. Domain module
- Extact FootballTeamRepository class to a separate file and moved it to domain module. Its implementations stay in app module
- Rename FootballTeamRepository.Impl class to RemoteFootballTeamRepository
- Create domain module where all domain specific classes should be. Moved FootballTeam, FootballTeamDetails, Gender and FootballTeamRepository classes to that module
