+++ app/src/main/java/ru/bykov/footballteams/di/Injection.kt
@@ -10,6 +10,7 @@ import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import retrofit2.converter.gson.GsonConverterFactory
 import ru.bykov.footballteams.models.FootballTeamRepository
 import ru.bykov.footballteams.models.InMemoryCachedFootballTeamRepository
+import ru.bykov.footballteams.models.RemoteFootballTeamRepository
 import ru.bykov.footballteams.network.TeamsApi
 
 const val BASE_URL = "https://android-exam.s3-eu-west-1.amazonaws.com"
@@ -51,7 +52,7 @@ object Injection {
 
     val repository: FootballTeamRepository by lazy {
         InMemoryCachedFootballTeamRepository(
-            FootballTeamRepository.Impl(teamsApi)
+            RemoteFootballTeamRepository(teamsApi)
         )
     }
 }
