+++ domain/src/main/kotlin/ru/bykov/footballteams/models/FootballTeamDetails.kt
