+++ domain/src/main/kotlin/ru/bykov/footballteams/models/FootballTeamRepository.kt
@@ -1,0 +1,11 @@
+package ru.bykov.footballteams.models
+
+import io.reactivex.Single
+
+interface FootballTeamRepository {
+
+    fun teams(forceUpdate: Boolean = false): Single<List<FootballTeam>>
+
+    fun details(id: Int): Single<FootballTeamDetails>
+
+}
