+++ gradle/platform/build.gradle.kts
@@ -21,6 +21,7 @@ dependencies.constraints {
     api("com.squareup.retrofit2:adapter-rxjava2:2.9.0")
     api("com.squareup.retrofit2:converter-gson:2.9.0")
     api("com.squareup.okhttp3:logging-interceptor:4.9.1")
+    api("com.google.code.gson:gson:2.10")
 
     // UI
     api("androidx.constraintlayout:constraintlayout:2.1.1")
